Improve build/rebuild perf via layer caching
In this PR, the packages layer from a previous build is made available
to the build proces thus speeding up the build, and later "pipenv clean"
is run to remove any packages that may have been removed in the rebuild.

There 2 cases of apps that use pipenv installation - Those with
Pipfile.lock checked in (1), and those without (2).

For case 2, neither a lock file is generated nor "clean" is run.
Generation of lock file is expensive and provides no real use here.
"Clean" is not run because it needs a lock file. This has a trade-off of
leaving packages not cleaned up in the images built from apps with no
lock file. Such users should clear their build cache or check-in a lock
file.

Signed-off-by: Arjun Sreedharan <asreedharan@vmware.com>